# contracts/openapi/afterchime-v1.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "Afterchime API",
    "version": "1.0.0",
    "description": "Development contract only; no production deployment is authorized."
  },
  "paths": {
    "/v1/installations": {
      "post": {
        "summary": "Register an installation and issue a session",
        "parameters": [
          { "in": "header", "name": "idempotency-key", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["appVersion", "device"],
                "properties": {
                  "appVersion": { "type": "string", "minLength": 1, "maxLength": 48, "pattern": "^[A-Za-z0-9.+_-]+$" },
                  "device": {
                    "type": "object",
                    "additionalProperties": false,
                    "required": ["androidApiLevel", "deviceClass"],
                    "properties": {
                      "androidApiLevel": { "type": "integer", "minimum": 26, "maximum": 100 },
                      "deviceClass": { "type": "string", "enum": ["phone", "tablet", "foldable"] }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Session" },
          "201": { "$ref": "#/components/responses/Session" },
          "409": { "description": "The idempotency key conflicts with a prior registration.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdentityError" } } } },
          "400": { "description": "The request does not match the strict registration schema." }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "summary": "Rotate a refresh-token family",
        "requestBody": { "$ref": "#/components/requestBodies/Refresh" },
        "responses": {
          "200": { "$ref": "#/components/responses/Session" },
          "401": { "$ref": "#/components/responses/InvalidRefresh" },
          "400": { "description": "The request does not match the strict refresh schema." }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "Revoke the refresh-token family; repeated logout is harmless",
        "requestBody": { "$ref": "#/components/requestBodies/Refresh" },
        "responses": {
          "204": { "description": "Refresh-token family revoked, or already absent." },
          "400": { "description": "The request does not match the strict logout schema." }
        }
      }
    },
    "/v1/analytics/events:batch": {
      "post": {
        "summary": "Ingest a bounded batch of consented analytics events",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "events"
                ],
                "properties": {
                  "events": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 50,
                    "items": {
                      "$ref": "../events/v1/envelope.schema.json"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Batch accepted for idempotent processing."
          },
          "400": {
            "description": "A schema or size check failed; rejected payload content is not retained."
          },
          "401": {
            "description": "Installation authentication failed."
          },
          "403": {
            "description": "Required consent is not active."
          },
          "413": {
            "description": "The batch exceeds the 128 KiB request limit."
          }
        }
      }
    },
    "/v1/notification-aggregates/daily": {
      "put": {
        "summary": "Idempotently replace one consented installation-owned daily notification aggregate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "category_counts",
                  "consent_scope_version",
                  "eligible_count",
                  "game_pressure",
                  "hourly_buckets",
                  "local_date",
                  "observation_completeness",
                  "revision",
                  "rules_version",
                  "timezone_offset_minutes"
                ],
                "properties": {
                  "category_counts": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "alarm": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "call": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "email": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "event": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "message": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "navigation": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "other": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "progress": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "reminder": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "social": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "transport": { "type": "integer", "minimum": 0, "maximum": 100000 },
                      "workout": { "type": "integer", "minimum": 0, "maximum": 100000 }
                    }
                  },
                  "consent_scope_version": { "type": "string", "minLength": 1, "maxLength": 32 },
                  "eligible_count": { "type": "integer", "minimum": 0, "maximum": 100000 },
                  "game_pressure": { "type": "integer", "minimum": 0, "maximum": 100 },
                  "hourly_buckets": {
                    "type": "array",
                    "maxItems": 24,
                    "items": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": ["count", "hour"],
                      "properties": {
                        "count": { "type": "integer", "minimum": 0, "maximum": 100000 },
                        "hour": { "type": "integer", "minimum": 0, "maximum": 23 }
                      }
                    }
                  },
                  "local_date": { "type": "string", "format": "date" },
                  "observation_completeness": { "type": "integer", "minimum": 0, "maximum": 100 },
                  "revision": { "type": "integer", "minimum": 1, "maximum": 2147483647 },
                  "rules_version": { "type": "integer", "minimum": 1, "maximum": 2147483647 },
                  "timezone_offset_minutes": { "type": "integer", "minimum": -840, "maximum": 840 }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Aggregate accepted or exact duplicate replayed."
          },
          "400": {
            "description": "Invalid, inconsistent, or unapproved aggregate payload; no aggregate is retained."
          },
          "401": {
            "description": "Installation authentication failed."
          },
          "403": {
            "description": "Notification aggregate consent is not active at the supplied version."
          },
          "409": {
            "description": "The daily aggregate already has the supplied or a newer conflicting revision."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "required": ["accessToken", "expiresInSeconds", "installationId", "refreshToken", "tokenType"],
        "properties": {
          "accessToken": { "type": "string" },
          "expiresInSeconds": { "type": "integer", "const": 900 },
          "installationId": { "type": "string", "format": "uuid" },
          "refreshToken": { "type": "string", "minLength": 43, "maxLength": 43, "pattern": "^[A-Za-z0-9_-]+$" },
          "tokenType": { "type": "string", "const": "Bearer" }
        }
      },
      "IdentityError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": { "error": { "type": "string", "enum": ["idempotency_conflict", "invalid_refresh_token"] } }
      }
    },
    "requestBodies": {
      "Refresh": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": false,
              "required": ["refreshToken"],
              "properties": { "refreshToken": { "type": "string", "minLength": 43, "maxLength": 43, "pattern": "^[A-Za-z0-9_-]+$" } }
            }
          }
        }
      }
    },
    "responses": {
      "Session": {
        "description": "A session was issued or rotated.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } }
      },
      "InvalidRefresh": {
        "description": "The refresh token is invalid, expired, revoked, or replayed; the family is revoked on replay.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdentityError" } } }
      }
    }
  }
}